Move login modal and modal to modals folder
